Move parking lot layer code into ParkingLotLoader.ts
