+++ src/js/map.ts
@@ -1,0 +1,58 @@
+import { Control, Map, TileLayer } from "leaflet";
+import "leaflet/dist/leaflet.css";
+
+const MAX_ZOOM = 18;
+const BASE_LAYERS = {
+  "High contrast": new TileLayer(
+    "https://tiles.stadiamaps.com/tiles/stamen_toner_lite/{z}/{x}/{y}{r}.png",
+    {
+      attribution: `&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a>
+        &copy; <a href="https://openmaptiles.org/" target="_blank">OpenMapTiles</a>
+        &copy; <a href="https://www.stadiamaps.com/" target="_blank">Stadia Maps</a>
+        &copy; <a href="https://www.stamen.com/" target="_blank">Stamen Design</a>`,
+      subdomains: "abcd",
+      minZoom: 0,
+      maxZoom: MAX_ZOOM,
+    }
+  ),
+  "Google Maps": new TileLayer(
+    "https://mt1.google.com/vt/lyrs=s&x={x}&y={y}&z={z}",
+    {
+      attribution: `&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> &copy; Google Maps`,
+      maxZoom: MAX_ZOOM,
+    }
+  ),
+};
+
+export const STYLES = {
+  cities: {
+    fillColor: "##c84041",
+    color: "#c84041",
+    weight: 4,
+    fillOpacity: 0,
+  },
+  parkingLots: {
+    fillColor: "#FF0000",
+    color: "#FF0000",
+    weight: 1,
+    fillOpacity: 0.6,
+  },
+};
+
+/**
+ * Create the initial map object.
+ *
+ * This sets up Google Maps vs. High contrast, attribution, and zoom.
+ */
+export function createMap(): Map {
+  const map = new Map("map", {
+    layers: [BASE_LAYERS["High contrast"]],
+  });
+  map.attributionControl.setPrefix(
+    '<a href="https://parkingreform.org/support/">Parking Reform Network</a>'
+  );
+
+  new Control.Layers(BASE_LAYERS).addTo(map);
+  map.createPane("fixed", document.getElementById("map") || undefined);
+  return map;
+}
